azure-function/requirements.txt
altair==5.5.0
attrs==25.3.0
azure-core==1.33.0
azure-cosmos==4.9.0
azure-functions==1.23.0
azure-identity==1.21.0
azure-keyvault-secrets==4.9.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
msal==1.32.3
msal-extensions==1.3.1
narwhals==1.37.1
numpy==2.2.5
packaging==24.2
pandas==2.2.3
pillow==11.2.1
protobuf==6.30.2
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
smmap==5.0.2
streamlit==1.45.0
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.1.3
